[All users] Fixed ending prolong conflict
The "ending prolong" hacky fix, which allows you to extend the duration of each screen in the ending's "white text" section, was overwriting the exact same code in the ending dialogue pointers file.

The “ending prolong” hacky fix was removed, and the relevant section in the ending dialogue pointers file is now better commented.