[All users] Changed Ending's music loop count
The epilogue's music loops halfway through the track. Looping the song 4 times for the English text causes that second half to never play, no matter how many epilogues play.

But, with 3 loops, even with the largest number of epilogues playing, the song will go all the way to the end, and stop right before Sara's epiloque, which is the last one.