[All users] Replaced effectiveness icons to save bigrams
Removing effectiveness icons and replacing them with Echoes' Anti-[X] descriptions saves a number of bigrams large enough (9) that this was worth a commit.

These icons served their purpose when I was running out of line space because of the two lines weapon description limit, but that time is long-gone.